Rebase controller-dev onto master
Add 1-axis mode


Test with 1-axis setup

Not yet tested
Get basic P control for 1 motor

Really janky, has saturation and slow settling time
Add defines for motor speed and saturation


Setup up timer 16 for use in motor commutation

Counter 16 interrupt defers to motor task in microseconds.
TIM16 only counts up so the CNT register value must be set to (0xffff - desiredDelay) in MICROseconds
Revert "Setup up timer 16 for use in motor commutation"

This reverts commit 149dd19bdbcdea493428ee54017efbe983919bed.

Use speed calculation mapping instead of delay mapping


Using microsecond timer, unstable


Fix Jitter issue

Make lowest possible delay 0xffff (and represent delay as uint32_t to make the math work out).
Reintroduce saturating the speed.
Make lowest speed correspond to 0xffff delay (2.04 degrees/s).
Increase maximum speed using microsecond timer

Gets rid of saturation zones for reasonable speeds
Add pulse curve script


Implement pulse-curving

Defines allow us to change the pulse params and the function defines the curve. This can be tuned later.
Generally, to move a load through a faster speed will require lots of torque, meaning higher pulse.
Sinusoidal target setting with PD control (untuned)

Tested on gimbal
Change calculation of euler angle implementation

Gives angles in +/- 180. Correct incorrect scaling of pitch.
Source: https://en.wikipedia.org/wiki/Conversion_between_quaternions_and_Euler_angles
Poor 1-axis gimballing with hardfault


Make encoders run at Fast Mode +

(1 MHz)
Remove queue-dispatch for IMU and Control loop

Directly wake control loop from IMU ISR
Use overrite to speed queue

improves performance at low speed
Implement 1-axis +/-80.0 deg limit


Add LED indication of IMU error for visualization

+++ BLDC_Camera_Gimbal_Firmware_STM32/Inc/encoder.h
@@ -32,7 +32,7 @@ typedef struct Encoder_t
 	Encoder_Identity_t identity;
 	I2C_HandleTypeDef* i2c;
 
-	uint16_t zeroPosition;
+	float zeroPosition;
 
 	uint16_t A1A2PinConfig;
 	uint16_t address;
@@ -49,6 +49,6 @@ typedef Encoder_t* Encoder_Handle_t;
 
 void Encoder_Init(Encoder_Handle_t encoder, Encoder_Identity_t identity, uint16_t pinconfig);
 float Poll_Encoder(Encoder_Handle_t encoder);
-void Set_Zero_Position(Encoder_Handle_t encoder, uint16_t zeropos);
+void Set_Zero_Position(Encoder_Handle_t encoder, float zeropos);
 
 #endif /* ENCODER_H_ */
